Extract InnerLessonList
