+++ src/pages/lessons/components/LessonList/InnerLessonList.tsx
@@ -1,0 +1,44 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+import type { LessonIndexEntry } from "types";
+import type { LessonListProps } from "pages/lessons/components/LessonList/LessonList";
+
+const WordCount = ({ lesson }: { lesson: LessonIndexEntry }) => (
+  <>{lesson.wordCount > 0 && ` · ${lesson.wordCount} words`}</>
+);
+
+const LessonLink = ({
+  lesson,
+  url,
+}: {
+  lesson: LessonIndexEntry;
+  url: string;
+}) => (
+  <Link
+    to={`${url}${lesson.path}`
+      .replace(/lesson\.txt$/, "")
+      .replace(/\/{2,}/g, "/")}
+    id={
+      "ga--lesson-index-" +
+      lesson.path.replace(/\/lesson\.txt/g, "").replace(/[/.]/g, "-")
+    }
+  >
+    {lesson.title}
+    {lesson.subtitle?.length > 0 && `: ${lesson.subtitle}`}
+  </Link>
+);
+
+export const InnerLessonList = ({
+  lessonIndex,
+  url,
+}: LessonListProps & { lessonIndex: LessonIndexEntry[] }) => (
+  <ul className="unstyled-list">
+    {lessonIndex.map((lesson) => (
+      <li className="unstyled-list-item mb1" key={lesson.path}>
+        <LessonLink lesson={lesson} url={url} />
+        <WordCount lesson={lesson} />
+      </li>
+    ))}
+  </ul>
+);
