+++ src/pages/lessons/components/LessonList/LessonList.tsx
@@ -5,50 +5,11 @@ import { groups } from "d3-array";
 import { useLessonIndex } from "../../../../states/lessonIndexState";
 import debounce from "../../../../utils/debounce";
 import filterLessons from "pages/lessons/components/LessonList/filterLessonsBySearchTerm";
-import type { LessonIndexEntry } from "../../../../types";
+import { InnerLessonList } from "pages/lessons/components/LessonList/InnerLessonList";
 
-type LessonListProps = {
+export type LessonListProps = {
   url: string;
 };
-
-const WordCount = ({ lesson }: { lesson: LessonIndexEntry }) => (
-  <>{lesson.wordCount > 0 && ` · ${lesson.wordCount} words`}</>
-);
-
-const LessonLink = ({
-  lesson,
-  url,
-}: {
-  lesson: LessonIndexEntry;
-  url: string;
-}) => (
-  <Link
-    to={`${url}${lesson.path}`
-      .replace(/lesson\.txt$/, "")
-      .replace(/\/{2,}/g, "/")}
-    id={
-      "ga--lesson-index-" +
-      lesson.path.replace(/\/lesson\.txt/g, "").replace(/[/.]/g, "-")
-    }
-  >
-    {lesson.title}
-    {lesson.subtitle?.length > 0 && `: ${lesson.subtitle}`}
-  </Link>
-);
-
-const InnerLessonList = ({
-  lessonIndex,
-  url,
-}: LessonListProps & { lessonIndex: LessonIndexEntry[] }) => (
-  <ul className="unstyled-list">
-    {lessonIndex.map((lesson) => (
-      <li className="unstyled-list-item mb1" key={lesson.path}>
-        <LessonLink lesson={lesson} url={url} />
-        <WordCount lesson={lesson} />
-      </li>
-    ))}
-  </ul>
-);
 
 const wrangleId = (id: string) => {
   return id.toLowerCase().replace(/[ ,’()]/g, "-");
